requirements.txt
absl-py==1.3.0
addict==2.4.0
albumentations==1.3.0
beautifulsoup4==4.11.1
cachetools==5.2.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==2.1.1
cmake==3.29.3
contourpy==1.0.7
cycler==0.11.0
Cython==3.0.10
filelock==3.14.0
fonttools==4.38.0
fsspec==2024.6.0
gdown==4.6.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
groundingdino-stk==1.0.0.2.0.118
grpcio==1.51.1
h5py==3.7.0
hickle==5.0.2
huggingface-hub==0.23.2
idna==3.4
imageio==2.24.0
importlib-resources==6.1.1
importlib_metadata==7.1.0
joblib==1.2.0
kiwisolver==1.4.4
lit==18.1.6
lvos==0.1.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.4
mpmath==1.3.0
networkx==3.1
numpy==1.23.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.68
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
platformdirs==4.2.2
progressbar2==4.2.0
protobuf==3.20.3
PuLP==2.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.7
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-utils==3.4.5
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
regex==2024.5.15
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
SimpleITK==2.3.1
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
spatial-correlation-sampler==0.4.0
supervision==0.6.0
sympy==1.12.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thinplate @ git+https://github.com/cheind/py-thin-plate-spline@f6995795397118b7d0ac01aecd3f39ffbfad9dee
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==1.0.3
tokenizers==0.19.1
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.4
transformers==4.41.2
triton==2.0.0
typing_extensions==4.12.1
urllib3==1.26.13
Werkzeug==2.2.2
yapf==0.40.2
zipp==3.11.0
